change internal resolution to 1080p
- Internal resolution is 1440 x 1080.
- Windowed resolutions are downscaled with antialiasing.
- Replaced assets with current project placeholders.
- Disabled mouse pointer.

+++ Source/Engine/GameServices.cs
@@ -28,7 +28,7 @@ namespace Engine
             RootScene.Enable();
         }
 
-        public void Initialize(int gameWidth, int gameHeight, int zoomMode, bool fullscreen)
+        public void Initialize(int gameWidth, int gameHeight, float zoomMode, bool fullscreen)
         {
             ResourceManager.Initialize();
             Error.StartLog();
